# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "MJ Ma",
            "url": "https://www.linkedin.com/in/mj-profile/",
            "email": "m.mjw.ma@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jwma/jump-jump/blob/master/LICENSE"
        }
    },
    "paths": {
...
        "/user/preferences": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户偏好设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "获取用户偏好设置",
                "responses": {
                    "200": {
                        "description": "data.preferences 为 []*models.UserPreference",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新用户偏好设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "更新用户偏好设置",
                "parameters": [
                    {
                        "description": "更新偏好设置请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data.preferences 为 []*models.UserPreference",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int",
                    "default": 0,
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "default": "ok",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}